+++ components/api-server/test/deletion.test.js
@@ -6,7 +6,7 @@
  */
 
 // @flow
-const nock = require('nock');
+const nock = require('nock'); 
 const cuid = require('cuid');
 const fs = require('fs');
 const path = require('path');
@@ -28,6 +28,9 @@ const {
 const { pubsub } = require('messages');
 const bluebird = require('bluebird');
 
+const cache = require('cache');
+const { MESSAGES } = require('cache/src/synchro');
+
 let app;
 let authKey;
 let username1; // fixtures reuse the username for userId
@@ -44,7 +47,7 @@ let config;
 let isOpenSource = false;
 let regUrl;
 
-describe('DELETE /users/:username', () => {
+describe('[PGTD] DELETE /users/:username', () => {
 
   before(async function() {
     config = await getConfig();
@@ -53,13 +56,18 @@ describe('DELETE /users/:username', () => {
     app = getApplication();
     await app.initiate();
 
-    await require('../src/methods/auth/delete')(app.api);
+    await require('api-server/src/methods/auth/delete')(app.api);
     let axonMsgs = [];
     const axonSocket = {
       emit: (...args) => axonMsgs.push(args),
     };
+    // needed even if not used
     pubsub.setTestNotifier(axonSocket);
     await require('api-server/src/methods/events')(app.api);
+    await require('api-server/src/methods/streams')(app.api);
+    await require('api-server/src/methods/auth/login')(app.api);
+    await require('api-server/src/methods/utility')(app.api);
+    await require('api-server/src/methods/auth/register')(app.api);
 
     request = supertest(app.expressApp);
 
@@ -132,9 +140,9 @@ describe('DELETE /users/:username', () => {
   // [isDnsLess, isOpenSource]
   const settingsToTest = [[true, false], [false, false], [true, true]];
   const testIDs = [
-    ['CM5Q', 'BQXA', '4Y76', '710F', 'GUPH', 'JNVS', 'C58U', 'IH6T', '75IW', 'MPXH'],
-    ['T21Z', 'K4J1', 'TIKT', 'WMMV', '9ZTM', 'T3UK', 'O73J', 'N8TR', '7WMG', 'UWYY'],
-    ['TPP2', '581Z', 'Z2FH', '4IH8', '33T6', 'SQ8P', '1F2Y', '7D0J', 'YD0B', 'L2Q1']];
+    ['CM5Q', 'BQXA', '4Y76', '710F', 'GUPH', 'JNVS', 'C58U', 'IH6T', '75IW', 'MPXH', '635G'],
+    ['T21Z', 'K4J1', 'TIKT', 'WMMV', '9ZTM', 'T3UK', 'O73J', 'N8TR', '7WMG', 'UWYY', 'U004'],
+    ['TPP2', '581Z', 'Z2FH', '4IH8', '33T6', 'SQ8P', '1F2Y', '7D0J', 'YD0B', 'L2Q1', 'CQ50']];
   for (let i = 0; i < settingsToTest.length; i++) {
     
     // skip tests that are not in scope
@@ -152,9 +160,10 @@ describe('DELETE /users/:username', () => {
       });
 
   
-      describe('when given existing username', function() {
+      describe('[D7HZ] when given existing username', function() {
         let deletedOnRegister = false;
         let userToDelete;
+        let natsDelivered = [];
         before(async function() {
 
           userToDelete = await initiateUserWithData(username1);
@@ -168,8 +177,15 @@ describe('DELETE /users/:username', () => {
             .times(1)
             .reply(200, { deleted: true });
           }
+          pubsub.setTestNatsDeliverHook(function(scopeName, eventName, payload) {
+            natsDelivered.push({scopeName, eventName, payload});
+          });
           res = await request.delete(`/users/${username1}`).set('Authorization', authKey);
         });
+        after(async function() {
+          pubsub.setTestNatsDeliverHook(null);
+        });
+
         it(`[${testIDs[i][0]}] should respond with 200`, function () {
           assert.equal(res.status, 200);
           assert.equal(res.body.userDeletion.username, username1);
@@ -219,6 +235,14 @@ describe('DELETE /users/:username', () => {
           const userFileExists = fs.existsSync(pathToUserAuditData);
           assert.isFalse(userFileExists);
         });
+        it(`[${testIDs[i][10]}] should delete user from the cache`, async function() {
+          const usersExists = cache.getUserId(userToDelete.attrs.id);
+          assert.isUndefined(usersExists);
+          assert.equal(natsDelivered.length, 1);
+          assert.equal(natsDelivered[0].scopeName, 'cache.' + MESSAGES.UNSET_USER);
+          assert.equal(natsDelivered[0].eventName, MESSAGES.UNSET_USER);
+          assert.equal(natsDelivered[0].payload.username, userToDelete.attrs.id);
+        })
         it(`[${testIDs[i][3]}] should not delete entries of other users`, async function() {
           const user = await usersRepository.getUserById(username2);
           assert.exists(user);
@@ -275,6 +299,72 @@ describe('DELETE /users/:username', () => {
       });
     });
   }
+  
+  describe('User - Create - Delete - Create - Login', function () { 
+    const usernamex = charlatan.Internet.userName().replace('_', '-') + 'x';
+
+    
+    it('[JBZM] should be able to recreate this user, and login', async function() {
+      nock(regUrl).post('/users/validate', () => {return true;}).times(2)
+      .reply(200, { errors: [] });
+
+      nock(regUrl).post('/users', () => {return true;}).times(2)
+      .reply(200, {username: usernamex});
+
+      async function createUser() {
+        const password = 'blupblipblop';
+        res = await request.post(`/users`)
+          .send({
+            appId: 'pryv-test',
+            username: usernamex,
+            password: 'blupblipblop',
+            email: usernamex + '@example.com',
+            insurancenumber: '123456789',
+          });
+
+        assert.equal(res.status, 201, 'should create a new user');
+        assert.isString(res.body.apiEndpoint, 'should receive an api Endpoint');
+        const token = res.body.apiEndpoint.split('//')[1].split('@')[0];
+
+        res = await request.post(`/${usernamex}/`).set('Authorization', token).send(
+          [{
+            method: 'streams.create', params: {id: 'diary', name: 'Journal'}
+          },
+          {
+            method: 'events.create', params: {streamId: 'diary', type: 'mass/kg', content: '70'}
+          }]
+        );
+        assert.equal(res.status, 200, 'should create a stream and an event');
+        assert.isArray(res.body.results, 'should receive an array of results');
+        assert.isObject(res.body.results[0].stream, 'should receive an stream');
+        assert.isObject(res.body.results[1].event, 'should receive an event');
+      };
+      
+      async function deleteUser() {
+        res = await request.delete(`/users/${usernamex}`).set('Authorization', authKey);
+        assert.equal(res.status, 200, 'should delete the user');
+        assert.equal(res.body.userDeletion?.username, usernamex, 'should receive the deleted username');
+      }
+
+      await createUser();
+      await deleteUser();
+      await createUser();
+      
+      res = await request.post('/' + usernamex + '/auth/login')
+      .set('Origin', 'http://test.pryv.local')
+      .send({
+        appId: 'pryv-test',
+        username: usernamex,
+        password: 'blupblipblop'
+      });
+      assert.equal(res.status, 200, 'should login');
+      assert.isString(res.body.apiEndpoint, 'should receive an api Endpoint');
+      assert.isString(res.body.token, 'should receive a token');
+
+      await deleteUser();
+    });
+  });
+
 });
 
 async function initiateUserWithData(username: string) {