+++ components/business/src/users/repository.js
@@ -94,7 +94,7 @@ class UsersRepository {
     return users;
   }
   async getUserIdForUsername(username: string) {
-    let userId = cache.get(cache.NS.USERID_BY_USERNAME, username);
+    let userId = cache.getUserId(username);
     if (! userId) {
       const userIdEvent = await bluebird.fromCallback(
         cb => this.eventsStorage.database.findOne(
@@ -116,7 +116,7 @@ class UsersRepository {
         ),
       );
       userId = userIdEvent?.userId;
-      cache.set(cache.NS.USERID_BY_USERNAME, username, userId);
+      cache.setUserId(username, userId);
     }
     return userId;
   }
@@ -147,9 +147,6 @@ class UsersRepository {
     let userId = await this.getUserIdForUsername(username);
     if (userId) {
       const user = await this.getUserById(userId);
-      if (! user) {
-         cache.unset(cache.NS.USERID_BY_USERNAME, username);
-      }
       return user;
     } 
     return null;
@@ -239,7 +236,6 @@ class UsersRepository {
     return true;
   }
   async insertOne(user: User, withSession: ?boolean = false): Promise<User> {
-    cache.unset(cache.NS.USERID_BY_USERNAME, user.username);
     // first explicitly create a collection, because it would fail in the transation
     await bluebird.fromCallback(
       cb => this.eventsStorage.database.getCollection(this.collectionInfo, cb),
@@ -287,8 +283,7 @@ class UsersRepository {
     return user;
   }
   async updateOne(user: User, update: {}, accessId: string): Promise<void> {
-    // invalidate caches
-    cache.unset(cache.NS.USERID_BY_USERNAME, user.username);
+   
    
     await this.checkDuplicates(update);
     
@@ -327,15 +322,21 @@ class UsersRepository {
       }
     }, getTransactionOptions());
   }
-  async deleteOne(userId: string): Promise<number> {
+  async deleteOne(userId: string, username: ?string): Promise<number> {
     const userAccountStreamsIds: Array<string> = SystemStreamsSerializer.getAccountStreamIds();
-    return await bluebird.fromCallback(
+    if (username == null) {
+      const user = await this.getUserById(userId);
+      username = user?.username;
+    }
+    const result = await bluebird.fromCallback(
       cb => this.eventsStorage.database.deleteMany(
         this.eventsStorage.getCollectionInfo(userId),
         { streamIds: { $in: userAccountStreamsIds } },
         cb,
       ),
     );
+    cache.unsetUser(username);
+    return result;
   }
   async checkUserPassword(userId: string, password: string): Promise<boolean> {
     const currentPass = await getUserPasswordHash(userId, this.eventsStorage);