+++ components/cache/src/index.js
@@ -4,15 +4,20 @@
  * Unauthorized copying of this file, via any medium is strictly prohibited
  * Proprietary and confidential
  */
+
+// @flow
 const { getLogger, getConfigUnsafe } = require('@pryv/boiler');
 const LRU = require('lru-cache');
 
+import type { Stream } from 'business/src/streams';
+
 const _caches = {};
-const MAX_PER_CACHE_SIZE = 2000; // maximum elements for each cache (namespace)
+const MAX_PER_CACHE_SIZE: number = 2000; // maximum elements for each cache (namespace)
 
 let synchro = null;
 
-let isActive = false;
+let isActive: boolean = false;
+let isSynchroActive: boolean = false
 
 const logger = getLogger('cache');
 const debug = {};
@@ -25,12 +30,17 @@ for (const key of ['set', 'get', 'unset', 'clear']) {
 
 const config = getConfigUnsafe(true);
 
-function getNameSpace(namespace) {
+/**
+ * username -> userId
+ */
+const userIdForUsername: Map<string, string> = new Map();
+
+function getNameSpace(namespace: string) {
   if (namespace == null) console.log('XXXX', new Error('Null namespace'));
   return _caches[namespace] || ( _caches[namespace] = new LRU(MAX_PER_CACHE_SIZE) ); // seting maxsize of 2000 to LRU cache
 }
 
-function set(namespace, key, value) {
+function set(namespace: string, key: string, value: string) {
   if (! isActive) return;  
   if (key == null) throw new Error('Null key for' + namespace);
   getNameSpace(namespace).set(key, value);
@@ -38,26 +48,28 @@ function set(namespace, key, value) {
   return value;
 }
 
-function unset(namespace, key) {
+function unset(namespace: string, key: string) {
   if (! isActive) return;
   if (key == null) throw new Error('Null key for' + namespace);
   getNameSpace(namespace).del(key);
   debug.unset(namespace, key);
 }
 
-function get(namespace, key) {
+function get(namespace: string, key: string) {
   if (! isActive) return null;
   if (key == null) throw new Error('Null key for' + namespace);
   debug.get(namespace, key);
   return getNameSpace(namespace).get(key);
 }
 
-function clear(namespace) {
+function clear(namespace: string) {
   if (namespace == null) { // clear all
     for (const ns of Object.keys(_caches)) {
       debug.clear(ns);
       delete _caches[ns];
     }
+    debug.clear('userIdForUsername');
+    userIdForUsername.clear();
   } else {
     delete _caches[namespace];
   }
@@ -65,43 +77,66 @@ function clear(namespace) {
   debug.clear(namespace);
 }
 
-function clearUserId(userId, notifyOtherProcesses = true) {
+//--------------- Users ---------------//
+
+function getUserId(username: string) {
+  return userIdForUsername.get(username);
+}
+
+function setUserId(username: string, userId: string) {
   if (! isActive) return;
-  if (notifyOtherProcesses && synchro != null) synchro.clearUserId(userId);
+  userIdForUsername.set(username, userId);
+}
+
+function unsetUser(username: string, notifyOtherProcesses: boolean = true) {
+  if (! isActive) return;
+  const userId = getUserId(username);
+  if (userId == null) return;
+  
+  unsetUserData(userId, false);
+  // notify userId delete
+  if (notifyOtherProcesses && isSynchroActive) synchro.unsetUser(userId);
+  userIdForUsername.delete(username);
+}
+
+function unsetUserData(userId: string, notifyOtherProcesses: boolean = true) {
+  if (! isActive) return;
+  // notify user data delete
+  if (notifyOtherProcesses && isSynchroActive) synchro.unsetUserData(userId);
   _unsetStreams(userId, 'local'); // for now we hardcode local streams
-  clearAccessLogics(userId);
+  _clearAccessLogics(userId);
 }
 
 //--------------- Streams ---------------//
-function getStreams(userId, key) {
-  return get(NS.STREAMS_FOR_USERID + key, userId);
+function getStreams(userId: string, storeId: string = 'local'): ?Array<Stream> {
+  return get(NS.STREAMS_FOR_USERID + storeId, userId);
 }
 
-function setStreams(userId, key, streams) {
+function setStreams(userId: string, storeId: string = 'local', streams: Array<Stream>): void {
   if (! isActive) return;
-  if (synchro != null) synchro.trackChangesForUserId(userId); // follow this user
-  set(NS.STREAMS_FOR_USERID + key, userId, streams);
+  if (isSynchroActive) synchro.registerListenerForUserId(userId); // follow this user
+  set(NS.STREAMS_FOR_USERID + storeId, userId, streams);
 }
 
-function _unsetStreams(userId, key) {
-  unset(NS.STREAMS_FOR_USERID + key, userId);
+function _unsetStreams(userId: string, storeId: string = 'local'): void {
+  unset(NS.STREAMS_FOR_USERID + storeId, userId);
 }
 
-function unsetStreams(userId, key) { 
-  clearUserId(userId); // for now we just fully clear this user.. 
+function unsetStreams(userId: string, storeId: string = 'local'): void { 
+  unsetUserData(userId);
 }
 
 
 //--------------- Access Logic -----------//
 
-function getAccessLogicForToken(userId, token) {
+function getAccessLogicForToken(userId: string, token: string) {
   if (! isActive) return null;
   const accessLogics = get(NS.ACCESS_LOGICS_FOR_USERID, userId);
   if (accessLogics == null) return null;
   return accessLogics.tokens[token];
 }
 
-function getAccessLogicForId(userId, accessId) {
+function getAccessLogicForId(userId: string, accessId: string) {
   if (! isActive) return null;
   const accessLogics = get(NS.ACCESS_LOGICS_FOR_USERID, userId);
   if (accessLogics == null) return null;
@@ -109,22 +144,24 @@ function getAccessLogicForId(userId, accessId) {
 }
 
 
-function unsetAccessLogic(userId, accessLogic, notifyOtherProcesses = true) {
+function unsetAccessLogic(userId: string, accessLogic: string, notifyOtherProcesses: boolean = true): void {
   if (! isActive) return;
-  if (notifyOtherProcesses && synchro != null) synchro.unsetAccessLogic(userId, accessLogic); // follow this user
+  // notify others to unsed
+  if (notifyOtherProcesses && isSynchroActive) synchro.unsetAccessLogic(userId, accessLogic); // follow this user
+  // perform unset
   const accessLogics = get(NS.ACCESS_LOGICS_FOR_USERID, userId);
   if (accessLogics == null) return ;
   delete accessLogics.tokens[accessLogic.token];
   delete accessLogics.ids[accessLogic.id];
 }
 
-function clearAccessLogics(userId) {
+function _clearAccessLogics(userId: string): void {
   unset(NS.ACCESS_LOGICS_FOR_USERID, userId);
 }
 
-function setAccessLogic(userId, accessLogic) {
+function setAccessLogic(userId: string, accessLogic: {}): void {
   if (! isActive) return;
-  if (synchro != null) synchro.trackChangesForUserId(userId);
+  if (synchro != null) synchro.registerListenerForUserId(userId);
   let accessLogics = get(NS.ACCESS_LOGICS_FOR_USERID, userId);
   if (accessLogics == null) {
     accessLogics = {
@@ -139,13 +176,6 @@ function setAccessLogic(userId, accessLogic) {
 
 //---------------
 
-/** Used only from tests to reload configuration after settting changes */
-function loadConfiguration() {
-  // could be true/false or 1/0 if launched from command line
-  isActive = config.get('caching:isActive') ? true : false;
-}
-loadConfiguration();
-
 const NS = {
   USERID_BY_USERNAME: 'USERID_BY_USERNAME',
   STREAMS_FOR_USERID: 'STREAMS',
@@ -153,33 +183,39 @@ const NS = {
 }
 
 const cache = {
-  set,
-  unset,
-  get,
   clear,
 
-  clearUserId,
+  getUserId,
+  setUserId,
+  unsetUser,
+  unsetUserData,
+
+  setStreams,
+  getStreams,
+  unsetStreams,
 
   getAccessLogicForId,
   getAccessLogicForToken,
   unsetAccessLogic,
   setAccessLogic,
 
-  setStreams,
-  getStreams,
-  unsetStreams,
-
   loadConfiguration,
   isActive,
 
-  NS
-}
+  NS,
+};
 
+/** Used only from tests to reload configuration after settting changes */
+function loadConfiguration() {
+  // could be true/false or 1/0 if launched from command line
+  isActive = config.get('caching:isActive') ? true : false;
+  isSynchroActive = config.get('openSource:isActive') ? false : true;
 
-// load synchro if needed
-if (! config.get('openSource:isActive')) {
-  synchro = require('./synchro.js');
-  synchro.setCache(cache);
+  if (isSynchroActive) {
+    synchro = require('./synchro.js');
+    synchro.setCache(cache);
+  }
 }
+loadConfiguration();
 
 module.exports = cache;