+++ components/messages/test/remover.test.js
@@ -4,6 +4,8 @@
  * Unauthorized copying of this file, via any medium is strictly prohibited
  * Proprietary and confidential
  */
+
+require('test-helpers/src/api-server-tests-config');
 const { pubsub } = require('messages');
 
 const { assert } = require('chai');