+++ components/storage/src/user/Streams.js
@@ -100,7 +100,7 @@ Streams.prototype.countAll = function (user, callback) {
 };
 
 Streams.prototype.insertOne = function (user, stream, callback) {
-  cache.clearUserId(user.id);
+  cache.unsetUserData(user.id);
   async.series([
     function checkDeletionWithSameId(stepDone) {
       if (! stream.id) { return stepDone(); }
@@ -123,7 +123,7 @@ Streams.prototype.insertOne = function (user, stream, callback) {
 
 Streams.prototype.updateOne = function (user, query, updatedData, callback) {
   if (typeof updatedData.parentId != 'undefined') { // clear ALL when a stream is moved
-    cache.clearUserId(user.id);
+    cache.unsetUserData(user.id);
   } else { // only stream Structure
     cache.unsetStreams(user.id, 'local');
   }
@@ -161,7 +161,7 @@ function checkParentExists(user, parentId, callback) {
  */
 Streams.prototype.delete = function (userOrUserId, query, callback) {
   const userId = userOrUserId.id || userOrUserId;
-  cache.clearUserId(userId);
+  cache.unsetUserData(userId);
   var update = {
     $set: {deleted: Date.now() / 1000},
     $unset: {